server/requirements.txt
absl-py
flask
json
pytest
typing